MapEditor : Tile.data finished + JComboBox for Map Name
The data features work, if a tile can be found in several versions, do a
right-click on it in the Tiles menu, it'll show a different version
everytime.
I added a JcomboBox instead of a JTextField in the EntityValues
interface to have it more secured (like the Cutscene Name thing I did
before).